Add flags parsing to nkv-server and nkv-client
So that user can specify socket to connect to and
directory for nkv-server where it will store files.
Instead of just specifying them in orderly manner I decided to
use flags, because it is more straightforward approach and it is backward
and forward-compatible friendly. In order to keep binary size as small as
possible, instead of brining a ready library which would increase size for 30-50Kbytes,
I wrote simple FlagParser, which does it's job and don't bring any 3rd party dependecies,
apart from HashMap and HashSet from collections, which were already part of nkv.

Signed-off-by: Pavel Abramov <uncle.decart@gmail.com>